+++ modules/iam-assumable-role/main.tf
@@ -50,6 +50,7 @@ data "aws_iam_policy_document" "assume_role" {
 
     dynamic "condition" {
       for_each = length(local.role_sts_externalid) != 0 ? [true] : []
+
       content {
         test     = "StringEquals"
         variable = "sts:ExternalId"
@@ -59,10 +60,21 @@ data "aws_iam_policy_document" "assume_role" {
 
     dynamic "condition" {
       for_each = var.role_requires_session_name ? [1] : []
+
       content {
         test     = "StringEquals"
         variable = "sts:RoleSessionName"
         values   = var.role_session_name
+      }
+    }
+
+    dynamic "condition" {
+      for_each = var.trust_policy_conditions
+
+      content {
+        test     = condition.value.test
+        variable = condition.value.variable
+        values   = condition.value.values
       }
     }
   }
@@ -121,6 +133,7 @@ data "aws_iam_policy_document" "assume_role_with_mfa" {
 
     dynamic "condition" {
       for_each = length(local.role_sts_externalid) != 0 ? [true] : []
+
       content {
         test     = "StringEquals"
         variable = "sts:ExternalId"
@@ -130,10 +143,21 @@ data "aws_iam_policy_document" "assume_role_with_mfa" {
 
     dynamic "condition" {
       for_each = var.role_requires_session_name ? [1] : []
+
       content {
         test     = "StringEquals"
         variable = "sts:RoleSessionName"
         values   = var.role_session_name
+      }
+    }
+
+    dynamic "condition" {
+      for_each = var.trust_policy_conditions
+
+      content {
+        test     = condition.value.test
+        variable = condition.value.variable
+        values   = condition.value.values
       }
     }
   }
@@ -196,4 +220,65 @@ resource "aws_iam_instance_profile" "this" {
   role  = aws_iam_role.this[0].name
 
   tags = var.tags
+}
+
+###############################
+# IAM Role Inline policy
+###############################
+
+locals {
+  create_iam_role_inline_policy = var.create_role && length(var.inline_policy_statements) > 0
+}
+
+data "aws_iam_policy_document" "inline" {
+  count = local.create_iam_role_inline_policy ? 1 : 0
+
+  dynamic "statement" {
+    for_each = var.inline_policy_statements
+
+    content {
+      sid           = try(statement.value.sid, null)
+      actions       = try(statement.value.actions, null)
+      not_actions   = try(statement.value.not_actions, null)
+      effect        = try(statement.value.effect, null)
+      resources     = try(statement.value.resources, null)
+      not_resources = try(statement.value.not_resources, null)
+
+      dynamic "principals" {
+        for_each = try(statement.value.principals, [])
+
+        content {
+          type        = principals.value.type
+          identifiers = principals.value.identifiers
+        }
+      }
+
+      dynamic "not_principals" {
+        for_each = try(statement.value.not_principals, [])
+
+        content {
+          type        = not_principals.value.type
+          identifiers = not_principals.value.identifiers
+        }
+      }
+
+      dynamic "condition" {
+        for_each = try(statement.value.conditions, [])
+
+        content {
+          test     = condition.value.test
+          values   = condition.value.values
+          variable = condition.value.variable
+        }
+      }
+    }
+  }
+}
+
+resource "aws_iam_role_policy" "inline" {
+  count = local.create_iam_role_inline_policy ? 1 : 0
+
+  role        = aws_iam_role.this[0].name
+  name_prefix = "${try(coalesce(var.role_name, var.role_name_prefix), "")}_inline_"
+  policy      = data.aws_iam_policy_document.inline[0].json
 }